+++ CHANGES.md
@@ -23,6 +23,7 @@
 * Add more functionality to add neural network functions to domain and use in parallel
 * Add special case of training heuristic with random policy when using pathfinding algorithm that uses a policy and not training a policy
 * Add timing of getting supervised data to supervised pathfinding
+* Add abstract method for loss info for training policy
 
 ## 0.2.1
 * Consolidate search: Beam search -> special cases: greedy_policy, graph search -> special cases: batch weighted A* search, batch weighted Q* search